chore: organize and move away font data into include/fonts
